LPS-70648 Moves code to AssetTagLocalServiceImpl
