+++ portal-impl/src/com/liferay/portlet/asset/service/impl/AssetTagLocalServiceImpl.java
@@ -21,16 +21,28 @@ import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCachable;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.SystemEventConstants;
 import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.search.BaseModelSearchResult;
+import com.liferay.portal.kernel.search.Document;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.Hits;
 import com.liferay.portal.kernel.search.Indexable;
 import com.liferay.portal.kernel.search.IndexableType;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
+import com.liferay.portal.kernel.search.QueryConfig;
+import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.systemevent.SystemEvent;
 import com.liferay.portal.kernel.util.ArrayUtil;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -39,9 +51,13 @@ import com.liferay.portlet.asset.util.AssetUtil;
 import com.liferay.portlet.asset.util.comparator.AssetTagNameComparator;
 import com.liferay.social.kernel.util.SocialCounterPeriodUtil;
 
+import java.io.Serializable;
+
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Provides the local service for accessing, adding, checking, deleting,
@@ -640,6 +656,114 @@ public class AssetTagLocalServiceImpl extends AssetTagLocalServiceBaseImpl {
 			groupIds, name, start, end, new AssetTagNameComparator());
 	}
 
+	@Override
+	public int searchCount(long groupId, String name) {
+		Indexer<AssetTag> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
+			AssetTag.class);
+
+		SearchContext searchContext = new SearchContext();
+
+		Map<String, Serializable> attributes = new HashMap<>();
+
+		attributes.put(Field.NAME, name);
+
+		searchContext.setAttributes(attributes);
+
+		searchContext.setEnd(QueryUtil.ALL_POS);
+		searchContext.setGroupIds(new long[] {groupId});
+		searchContext.setKeywords(name);
+		searchContext.setStart(QueryUtil.ALL_POS);
+
+		QueryConfig queryConfig = searchContext.getQueryConfig();
+
+		queryConfig.setHighlightEnabled(false);
+		queryConfig.setScoreEnabled(false);
+
+		for (int i = 0; i < 10; i++) {
+			Hits hits = null;
+
+			try {
+				hits = indexer.search(searchContext);
+			}
+			catch (SearchException se) {
+				_log.error("Unable to find asset tag documents", se);
+
+				continue;
+			}
+
+			List<AssetTag> tags = getTags(hits);
+
+			if (tags != null) {
+				BaseModelSearchResult<AssetTag> baseModelSearchResult =
+					new BaseModelSearchResult<>(tags, hits.getLength());
+
+				return baseModelSearchResult.getLength();
+			}
+		}
+	}
+
+
+	@Override
+	public List<AssetTag> searchTags(
+		long[] groupIds, String name, int start, int end) {
+
+		Indexer<AssetTag> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
+			AssetTag.class);
+
+		SearchContext searchContext = new SearchContext();
+
+		ServiceContext serviceContext =
+			ServiceContextThreadLocal.getServiceContext();
+
+		if (serviceContext != null) {
+			searchContext.setCompanyId(serviceContext.getCompanyId());
+			searchContext.setUserId(serviceContext.getUserId());
+		}
+
+		name = HtmlUtil.escapeURL(name);
+
+		Map<String, Serializable> attributes = new HashMap<>();
+
+		attributes.put(Field.NAME, name);
+
+		searchContext.setAttributes(attributes);
+
+		searchContext.setEnd(end);
+		searchContext.setGroupIds(groupIds);
+		searchContext.setKeywords(name);
+		searchContext.setStart(start);
+
+		QueryConfig queryConfig = searchContext.getQueryConfig();
+
+		queryConfig.setHighlightEnabled(false);
+		queryConfig.setScoreEnabled(false);
+
+		for (int i = 0; i < 10; i++) {
+			Hits hits = null;
+
+			try {
+				hits = indexer.search(searchContext);
+			}
+			catch (SearchException se) {
+				_log.error("Unable to find asset tag documents", se);
+
+				continue;
+			}
+
+			List<AssetTag> tags = getTags(hits);
+
+			if (tags != null) {
+				BaseModelSearchResult<AssetTag> baseModelSearchResult =
+					new BaseModelSearchResult<>(tags, hits.getLength());
+
+				return baseModelSearchResult.getBaseModels();
+			}
+		}
+
+		return assetTagPersistence.findByG_LikeN(
+			groupIds, name, start, end, new AssetTagNameComparator());
+	}
+
 	@Indexable(type = IndexableType.REINDEX)
 	@Override
 	public AssetTag updateTag(
@@ -692,6 +816,43 @@ public class AssetTagLocalServiceImpl extends AssetTagLocalServiceBaseImpl {
 		return ListUtil.toArray(tags, AssetTag.NAME_ACCESSOR);
 	}
 
+	protected List<AssetTag> getTags(Hits hits) {
+		List<Document> documents = hits.toList();
+
+		List<AssetTag> tags = new ArrayList<>(documents.size());
+
+		for (Document document : documents) {
+			long tagId = GetterUtil.getLong(document.get(Field.ENTRY_CLASS_PK));
+
+			AssetTag tag = assetTagLocalService.fetchAssetTag(tagId);
+
+			if (tag == null) {
+				tags = null;
+
+				Indexer<AssetTag> indexer = IndexerRegistryUtil.getIndexer(
+					AssetTag.class);
+
+				long companyId = GetterUtil.getLong(
+					document.get(Field.COMPANY_ID));
+
+				try {
+					indexer.delete(companyId, document.getUID());
+				}
+				catch (SearchException se) {
+					_log.error(
+						"Unable to delete asset tag document" +
+							document.getUID(),
+						se);
+				}
+			}
+			else if (tags != null) {
+				tags.add(tag);
+			}
+		}
+
+		return tags;
+	}
+
 	protected void validate(String name) throws PortalException {
 		if (!AssetUtil.isValidWord(name)) {
 			throw new AssetTagException(
@@ -700,5 +861,8 @@ public class AssetTagLocalServiceImpl extends AssetTagLocalServiceBaseImpl {
 				AssetTagException.INVALID_CHARACTER);
 		}
 	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		AssetTagLocalServiceImpl.class);
 
 }
