+++ portal-impl/src/com/liferay/portlet/asset/service/impl/AssetTagServiceImpl.java
@@ -19,26 +19,9 @@ import com.liferay.asset.kernel.model.AssetTagDisplay;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.BaseModelSearchResult;
-import com.liferay.portal.kernel.search.Document;
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.Hits;
-import com.liferay.portal.kernel.search.Indexer;
-import com.liferay.portal.kernel.search.IndexerRegistryUtil;
-import com.liferay.portal.kernel.search.QueryConfig;
-import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.HtmlUtil;
-import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portlet.asset.service.base.AssetTagServiceBaseImpl;
 import com.liferay.portlet.asset.service.permission.AssetTagPermission;
@@ -47,12 +30,8 @@ import com.liferay.portlet.asset.util.comparator.AssetTagNameComparator;
 import com.liferay.util.Autocomplete;
 import com.liferay.util.dao.orm.CustomSQLUtil;
 
-import java.io.Serializable;
-
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 
@@ -200,66 +179,11 @@ public class AssetTagServiceImpl extends AssetTagServiceBaseImpl {
 		OrderByComparator<AssetTag> obc) {
 
 		if (Validator.isNull(name)) {
-			return assetTagPersistence.findByGroupId(groupIds, start, end);
+			return assetTagPersistence.findByGroupId(groupIds, start, end, obc);
 		}
-
-		Indexer<AssetTag> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			AssetTag.class);
-
-		SearchContext searchContext = new SearchContext();
-
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			searchContext.setCompanyId(serviceContext.getCompanyId());
-			searchContext.setUserId(serviceContext.getUserId());
-		}
-
-		name = HtmlUtil.escapeURL(name);
-
-		Map<String, Serializable> attributes = new HashMap<>();
-
-		attributes.put(Field.NAME, name);
-
-		searchContext.setAttributes(attributes);
-
-		searchContext.setEnd(end);
-		searchContext.setGroupIds(groupIds);
-		searchContext.setKeywords(name);
-		searchContext.setStart(start);
-
-		QueryConfig queryConfig = searchContext.getQueryConfig();
-
-		queryConfig.setHighlightEnabled(false);
-		queryConfig.setScoreEnabled(false);
-
-		for (int i = 0; i < 10; i++) {
-			Hits hits = null;
-
-			try {
-				hits = indexer.search(searchContext);
-			}
-			catch (SearchException se) {
-				_log.error("Unable to find asset tag documents", se);
-
-				continue;
-			}
-
-			List<AssetTag> tags = getTags(hits);
-
-			if (tags != null) {
-				BaseModelSearchResult<AssetTag> baseModelSearchResult =
-					new BaseModelSearchResult<>(tags, hits.getLength());
-
-				return baseModelSearchResult.getBaseModels();
-			}
-		}
-
-		name = HttpUtil.decodeURL(StringUtil.quote(name, StringPool.PERCENT));
 
 		return assetTagPersistence.findByG_LikeN(
-			groupIds, name, start, end, new AssetTagNameComparator());
+			groupIds, name, start, end, obc);
 	}
 
 	@Override
@@ -271,49 +195,6 @@ public class AssetTagServiceImpl extends AssetTagServiceBaseImpl {
 	public int getTagsCount(long groupId, String name) {
 		if (Validator.isNull(name)) {
 			return assetTagPersistence.countByGroupId(groupId);
-		}
-
-		Indexer<AssetTag> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			AssetTag.class);
-
-		SearchContext searchContext = new SearchContext();
-
-		Map<String, Serializable> attributes = new HashMap<>();
-
-		attributes.put(Field.NAME, name);
-
-		searchContext.setAttributes(attributes);
-
-		searchContext.setEnd(QueryUtil.ALL_POS);
-		searchContext.setGroupIds(new long[] {groupId});
-		searchContext.setKeywords(name);
-		searchContext.setStart(QueryUtil.ALL_POS);
-
-		QueryConfig queryConfig = searchContext.getQueryConfig();
-
-		queryConfig.setHighlightEnabled(false);
-		queryConfig.setScoreEnabled(false);
-
-		for (int i = 0; i < 10; i++) {
-			Hits hits = null;
-
-			try {
-				hits = indexer.search(searchContext);
-			}
-			catch (SearchException se) {
-				_log.error("Unable to find asset tag documents", se);
-
-				continue;
-			}
-
-			List<AssetTag> tags = getTags(hits);
-
-			if (tags != null) {
-				BaseModelSearchResult<AssetTag> baseModelSearchResult =
-					new BaseModelSearchResult<>(tags, hits.getLength());
-
-				return baseModelSearchResult.getLength();
-			}
 		}
 
 		return assetTagPersistence.countByG_LikeN(groupId, name);
@@ -371,45 +252,5 @@ public class AssetTagServiceImpl extends AssetTagServiceBaseImpl {
 		return assetTagLocalService.updateTag(
 			getUserId(), tagId, name, serviceContext);
 	}
-
-	protected List<AssetTag> getTags(Hits hits) {
-		List<Document> documents = hits.toList();
-
-		List<AssetTag> tags = new ArrayList<>(documents.size());
-
-		for (Document document : documents) {
-			long tagId = GetterUtil.getLong(document.get(Field.ENTRY_CLASS_PK));
-
-			AssetTag tag = assetTagLocalService.fetchAssetTag(tagId);
-
-			if (tag == null) {
-				tags = null;
-
-				Indexer<AssetTag> indexer = IndexerRegistryUtil.getIndexer(
-					AssetTag.class);
-
-				long companyId = GetterUtil.getLong(
-					document.get(Field.COMPANY_ID));
-
-				try {
-					indexer.delete(companyId, document.getUID());
-				}
-				catch (SearchException se) {
-					_log.error(
-						"Unable to delete asset tag document" +
-							document.getUID(),
-						se);
-				}
-			}
-			else if (tags != null) {
-				tags.add(tag);
-			}
-		}
-
-		return tags;
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		AssetTagServiceImpl.class);
 
 }
